refactor: refactor code structor using generator-eslint
