--- index.js
@@ -1,6 +1,0 @@
-const desturctoringFormState = require("./destructoringFormState");
-module.exports = {
-  rules: {
-    "destructuring-formState": desturctoringFormState,
-  },
-};
