+++ package.json
@@ -1,25 +1,28 @@
 {
   "name": "eslint-plugin-react-hook-form",
   "version": "0.1.0",
-  "description": "eslint plugin for react-hook-form",
-  "main": "index.js",
-  "scripts": {},
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/andykao1213/eslint-plugin-react-hook-form.git"
-  },
+  "description": "ESLint plugin for react-hook-form",
   "keywords": [
     "react",
+    "react-hook-form",
     "eslint",
-    "react-hook-form"
+    "eslintplugin",
+    "eslint-plugin"
   ],
   "author": "Andrew Kao",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/andykao1213/eslint-plugin-react-hook-form/issues"
+  "main": "lib/index.js",
+  "scripts": {
+    "test": "mocha tests --recursive"
   },
-  "homepage": "https://github.com/andykao1213/eslint-plugin-react-hook-form#readme",
+  "dependencies": {
+    "requireindex": "~1.1.0"
+  },
   "devDependencies": {
-    "eslint": "^7.28.0"
-  }
+    "eslint": "^7.1.0",
+    "mocha": "^7.2.0"
+  },
+  "engines": {
+    "node": ">=0.10.0"
+  },
+  "license": "MIT"
 }
